Replace deprecated calls to tf.to_(float|int32) with proper tf.cast() calls.
PiperOrigin-RevId: 251286951

+++ tensorflow_federated/python/learning/federated_evaluation_test.py
@@ -62,7 +62,8 @@ class TestModel(model.Model):
   def forward_pass(self, batch, training=True):
     assert not training
     num_over = tf.reduce_sum(
-        tf.to_float(tf.greater(batch['temp'], self._variables.max_temp)))
+        tf.cast(
+            tf.greater(batch['temp'], self._variables.max_temp), tf.float32))
     tf.assign_add(self._variables.num_over, num_over)
     loss = tf.constant(0.0)
     predictions = tf.zeros_like(batch['temp'])
